refactor: use import alias, move constants from util file
